Extract shared xstate E2E helpers, fix bookingLogs matching and unhandled promises
- Create tests/e2e/helpers/xstate-mocks.ts with shared constants and utilities
- Fix includes("booking") matching bookingLogs in all 5 xstate test files
- Add void/catch to fire-and-forget loadBookings() calls
- Remove duplicated helpers (createTimestamp, serialize*, TIMESTAMP_FIELDS, etc.)
- Remove unused Timestamp imports
