+++ booking-app/tests/e2e/xstate-cancel-flows.e2e.test.ts
@@ -8,11 +8,15 @@ import {
   doc,
   getDoc,
   setDoc,
-  Timestamp,
 } from "../../lib/firebase/stubs/firebaseFirestoreStub";
 import { registerBookingMocks } from "./helpers/mock-routes";
-
-const jsonHeaders = { "content-type": "application/json" };
+import {
+  JSON_HEADERS,
+  TIMESTAMP_FIELDS,
+  createTimestamp,
+  serializeBookingRecord,
+  serializeGenericRecord,
+} from "./helpers/xstate-mocks";
 
 const BASE_URL =
   process.env.PLAYWRIGHT_BASE_URL ?? "http://localhost:3000";
@@ -27,28 +31,6 @@ const USERS_RIGHTS_DOC_ID = "mock-cancel-rights";
 const USERS_APPROVER_DOC_ID = "mock-cancel-approver";
 
 const REQUEST_NUMBER = 9850;
-
-const TIMESTAMP_FIELDS = [
-  "startDate",
-  "endDate",
-  "requestedAt",
-  "firstApprovedAt",
-  "finalApprovedAt",
-  "declinedAt",
-  "canceledAt",
-  "checkedInAt",
-  "checkedOutAt",
-  "noShowedAt",
-  "closedAt",
-  "walkedInAt",
-];
-
-const createTimestamp = (date: Date) => {
-  const ts = new Timestamp(date);
-  (ts as any).toMillis = () => date.getTime();
-  (ts as any).toJSON = () => date.toISOString();
-  return ts;
-};
 
 async function seedLiaisonUserData() {
   const now = createTimestamp(new Date());
@@ -202,45 +184,6 @@ async function seedBooking(opts: {
   });
 }
 
-function serializeBookingRecord(record: any) {
-  const serialized: Record<string, any> = { ...record };
-
-  for (const field of TIMESTAMP_FIELDS) {
-    const value = record[field];
-    if (value && typeof value.toDate === "function") {
-      serialized[field] = value.toDate().toISOString();
-    }
-  }
-
-  if (record.startDate?.toDate) {
-    serialized.startDate = record.startDate.toDate().toISOString();
-  }
-  if (record.endDate?.toDate) {
-    serialized.endDate = record.endDate.toDate().toISOString();
-  }
-  if (record.requestedAt?.toDate) {
-    serialized.requestedAt = record.requestedAt.toDate().toISOString();
-  }
-
-  serialized.xstateData = record.xstateData
-    ? JSON.parse(JSON.stringify(record.xstateData))
-    : undefined;
-
-  return serialized;
-}
-
-function serializeGenericRecord(record: any) {
-  const serialized: Record<string, any> = { ...record };
-
-  Object.entries(serialized).forEach(([key, value]) => {
-    if (value && typeof (value as any).toDate === "function") {
-      serialized[key] = (value as any).toDate().toISOString();
-    }
-  });
-
-  return serialized;
-}
-
 async function registerMockBookingsFeed(
   page: Page,
   userRole: "liaison" | "admin"
@@ -294,7 +237,7 @@ async function registerMockBookingsFeed(
 
     await route.fulfill({
       status: 200,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify(payload),
     });
   });
@@ -405,7 +348,8 @@ async function registerMockBookingsFeed(
 
         if (
           normalizedTableName.includes("booking") &&
-          !normalizedTableName.includes("type")
+          !normalizedTableName.includes("type") &&
+          !normalizedTableName.includes("log")
         ) {
           return await ensureBookings();
         }
@@ -444,7 +388,8 @@ async function registerMockBookingsFeed(
         if (
           path &&
           path.includes("booking") &&
-          !path.includes("bookingTypes")
+          !path.includes("bookingTypes") &&
+          !path.includes("bookingLog")
         ) {
           const bookings = await ensureBookings();
           return {
@@ -593,7 +538,7 @@ async function mockTransitionEndpoints(page: Page) {
 
       await route.fulfill({
         status: 200,
-        headers: jsonHeaders,
+        headers: JSON_HEADERS,
         body: JSON.stringify({
           success: true,
           newState,
@@ -603,7 +548,7 @@ async function mockTransitionEndpoints(page: Page) {
     }
     await route.fulfill({
       status: 405,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify({ error: "Method Not Allowed" }),
     });
   });
@@ -612,7 +557,7 @@ async function mockTransitionEndpoints(page: Page) {
     if (route.request().method() === "POST") {
       await route.fulfill({
         status: 200,
-        headers: jsonHeaders,
+        headers: JSON_HEADERS,
         body: JSON.stringify({ success: true }),
       });
       return;
@@ -620,7 +565,7 @@ async function mockTransitionEndpoints(page: Page) {
 
     await route.fulfill({
       status: 200,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify([]),
     });
   });
