+++ booking-app/tests/e2e/xstate-service-closeout.e2e.test.ts
@@ -8,11 +8,16 @@ import {
   doc,
   getDoc,
   setDoc,
-  Timestamp,
 } from "../../lib/firebase/stubs/firebaseFirestoreStub";
 import { registerBookingMocks } from "./helpers/mock-routes";
-
-const jsonHeaders = { "content-type": "application/json" };
+import {
+  JSON_HEADERS,
+  TIMESTAMP_FIELDS_WITH_BY,
+  createTimestamp,
+  serializeBookingRecord,
+  serializeGenericRecord,
+  registerDefinePropertyInterceptor,
+} from "./helpers/xstate-mocks";
 
 const BASE_URL =
   process.env.PLAYWRIGHT_BASE_URL ?? "http://localhost:3000";
@@ -27,28 +32,6 @@ const USERS_RIGHTS_DOC_ID = "mock-closeout-rights";
 const USERS_APPROVER_DOC_ID = "mock-closeout-approver";
 
 const REQUEST_NUMBER = 9880;
-
-const TIMESTAMP_FIELDS = [
-  "startDate",
-  "endDate",
-  "requestedAt",
-  "firstApprovedAt",
-  "firstApprovedBy",
-  "finalApprovedAt",
-  "finalApprovedBy",
-  "declinedAt",
-  "declinedBy",
-  "canceledAt",
-  "canceledBy",
-  "checkedInAt",
-  "checkedInBy",
-  "checkedOutAt",
-  "checkedOutBy",
-  "noShowedAt",
-  "noShowedBy",
-  "closedAt",
-  "closedBy",
-];
 
 const SERVICE_METADATA = {
   staff: {
@@ -112,13 +95,6 @@ const SERVICE_METADATA = {
     closeoutDone: "Setup Closedout",
   },
 } as const;
-
-const createTimestamp = (date: Date) => {
-  const ts = new Timestamp(date);
-  (ts as any).toMillis = () => date.getTime();
-  (ts as any).toJSON = () => date.toISOString();
-  return ts;
-};
 
 async function seedAdminUserData() {
   const now = createTimestamp(new Date());
@@ -285,45 +261,6 @@ async function seedCloseoutBooking(
   await setDoc(doc({} as any, "mc-bookings", BOOKING_DOC_ID), bookingData);
 }
 
-function serializeBookingRecord(record: any) {
-  const serialized: Record<string, any> = { ...record };
-
-  for (const field of TIMESTAMP_FIELDS) {
-    const value = record[field];
-    if (value && typeof value.toDate === "function") {
-      serialized[field] = value.toDate().toISOString();
-    }
-  }
-
-  if (record.startDate?.toDate) {
-    serialized.startDate = record.startDate.toDate().toISOString();
-  }
-  if (record.endDate?.toDate) {
-    serialized.endDate = record.endDate.toDate().toISOString();
-  }
-  if (record.requestedAt?.toDate) {
-    serialized.requestedAt = record.requestedAt.toDate().toISOString();
-  }
-
-  serialized.xstateData = record.xstateData
-    ? JSON.parse(JSON.stringify(record.xstateData))
-    : undefined;
-
-  return serialized;
-}
-
-function serializeGenericRecord(record: any) {
-  const serialized: Record<string, any> = { ...record };
-
-  Object.entries(serialized).forEach(([key, value]) => {
-    if (value && typeof (value as any).toDate === "function") {
-      serialized[key] = (value as any).toDate().toISOString();
-    }
-  });
-
-  return serialized;
-}
-
 async function registerMockBookingsFeed(page: Page) {
   const snapshot = await getDoc(
     doc({} as any, "mc-bookings", BOOKING_DOC_ID)
@@ -364,36 +301,12 @@ async function registerMockBookingsFeed(page: Page) {
 
     await route.fulfill({
       status: 200,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify(payload),
     });
   });
 
-  await page.addInitScript(() => {
-    const targetExports = new Set([
-      "clientGetDataByCalendarEventId",
-      "clientFetchAllDataFromCollection",
-      "getPaginatedData",
-    ]);
-    const origDefineProperty = Object.defineProperty;
-    Object.defineProperty = function (
-      obj: any,
-      prop: PropertyKey,
-      descriptor: PropertyDescriptor
-    ) {
-      if (
-        descriptor &&
-        descriptor.get &&
-        !descriptor.configurable &&
-        typeof prop === "string" &&
-        targetExports.has(prop)
-      ) {
-        descriptor = { ...descriptor, configurable: true };
-      }
-      return origDefineProperty.call(this, obj, prop, descriptor);
-    } as typeof Object.defineProperty;
-    Object.defineProperty.toString = () => "function defineProperty() { [native code] }";
-  });
+  await registerDefinePropertyInterceptor(page);
 
   await page.addInitScript(
     ({
@@ -505,7 +418,8 @@ async function registerMockBookingsFeed(page: Page) {
 
         if (
           normalizedTableName.includes("booking") &&
-          !normalizedTableName.includes("type")
+          !normalizedTableName.includes("type") &&
+          !normalizedTableName.includes("log")
         ) {
           return await ensureBookings();
         }
@@ -544,7 +458,8 @@ async function registerMockBookingsFeed(page: Page) {
         if (
           path &&
           path.includes("booking") &&
-          !path.includes("bookingTypes")
+          !path.includes("bookingTypes") &&
+          !path.includes("bookingLog")
         ) {
           const bookings = await ensureBookings();
           return {
@@ -579,7 +494,7 @@ async function registerMockBookingsFeed(page: Page) {
         return { docs: [] };
       };
 
-      loadBookings();
+      void loadBookings().catch(() => {});
 
       const overrideClientGetDataById = async (calendarEventId: string) => {
         const bookings = await ensureBookings();
@@ -682,7 +597,7 @@ async function registerMockBookingsFeed(page: Page) {
       };
     },
     {
-      timestampFields: TIMESTAMP_FIELDS,
+      timestampFields: TIMESTAMP_FIELDS_WITH_BY,
       initialBooking: basePayload,
       initialUsersRights: rightsPayload,
       adminEmail: ADMIN_EMAIL,
@@ -741,7 +656,7 @@ async function mockServicesEndpoint(page: Page, captured: any[]) {
     if (route.request().method() !== "POST") {
       await route.fulfill({
         status: 405,
-        headers: jsonHeaders,
+        headers: JSON_HEADERS,
         body: JSON.stringify({ error: "Method Not Allowed" }),
       });
       return;
@@ -769,7 +684,7 @@ async function mockServicesEndpoint(page: Page, captured: any[]) {
 
     await route.fulfill({
       status: 200,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify({ success: true }),
     });
   });
@@ -780,14 +695,14 @@ async function mockTransitionEndpoints(page: Page) {
     if (route.request().method() === "POST") {
       await route.fulfill({
         status: 200,
-        headers: jsonHeaders,
+        headers: JSON_HEADERS,
         body: JSON.stringify({ success: true }),
       });
       return;
     }
     await route.fulfill({
       status: 405,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify({ error: "Method Not Allowed" }),
     });
   });
@@ -796,14 +711,14 @@ async function mockTransitionEndpoints(page: Page) {
     if (route.request().method() === "POST") {
       await route.fulfill({
         status: 200,
-        headers: jsonHeaders,
+        headers: JSON_HEADERS,
         body: JSON.stringify({ success: true }),
       });
       return;
     }
     await route.fulfill({
       status: 200,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify([]),
     });
   });
