+++ booking-app/tests/e2e/xstate-service-decline.e2e.test.ts
@@ -8,11 +8,15 @@ import {
   doc,
   getDoc,
   setDoc,
-  Timestamp,
 } from "../../lib/firebase/stubs/firebaseFirestoreStub";
 import { registerBookingMocks } from "./helpers/mock-routes";
-
-const jsonHeaders = { "content-type": "application/json" };
+import {
+  JSON_HEADERS,
+  TIMESTAMP_FIELDS_WITH_BY,
+  createTimestamp,
+  serializeBookingRecord,
+  registerDefinePropertyInterceptor,
+} from "./helpers/xstate-mocks";
 
 const BASE_URL =
   process.env.PLAYWRIGHT_BASE_URL ?? "http://localhost:3000";
@@ -29,28 +33,6 @@ const USERS_RIGHTS_DOC_ID = "mock-svc-decline-rights";
 const USERS_APPROVER_DOC_ID = "mock-svc-decline-approver";
 
 const REQUEST_NUMBER = 9870;
-
-const TIMESTAMP_FIELDS = [
-  "startDate",
-  "endDate",
-  "requestedAt",
-  "firstApprovedAt",
-  "firstApprovedBy",
-  "finalApprovedAt",
-  "finalApprovedBy",
-  "declinedAt",
-  "declinedBy",
-  "canceledAt",
-  "canceledBy",
-  "checkedInAt",
-  "checkedInBy",
-  "checkedOutAt",
-  "checkedOutBy",
-  "noShowedAt",
-  "noShowedBy",
-  "closedAt",
-  "closedBy",
-];
 
 const SERVICE_METADATA = {
   staff: {
@@ -96,13 +78,6 @@ const SERVICE_METADATA = {
     declinedState: "Setup Declined",
   },
 } as const;
-
-const createTimestamp = (date: Date) => {
-  const ts = new Timestamp(date);
-  (ts as any).toMillis = () => date.getTime();
-  (ts as any).toJSON = () => date.toISOString();
-  return ts;
-};
 
 async function seedServicesUserData() {
   const now = createTimestamp(new Date());
@@ -362,33 +337,6 @@ async function seedNonServiceBooking() {
   });
 }
 
-function serializeBookingRecord(record: any) {
-  const serialized: Record<string, any> = { ...record };
-
-  for (const field of TIMESTAMP_FIELDS) {
-    const value = record[field];
-    if (value && typeof value.toDate === "function") {
-      serialized[field] = value.toDate().toISOString();
-    }
-  }
-
-  if (record.startDate?.toDate) {
-    serialized.startDate = record.startDate.toDate().toISOString();
-  }
-  if (record.endDate?.toDate) {
-    serialized.endDate = record.endDate.toDate().toISOString();
-  }
-  if (record.requestedAt?.toDate) {
-    serialized.requestedAt = record.requestedAt.toDate().toISOString();
-  }
-
-  serialized.xstateData = record.xstateData
-    ? JSON.parse(JSON.stringify(record.xstateData))
-    : undefined;
-
-  return serialized;
-}
-
 async function registerMockBookingsFeed(page: Page) {
   const docIds = [SERVICE_BOOKING_ID, NON_SERVICE_BOOKING_ID];
 
@@ -409,7 +357,7 @@ async function registerMockBookingsFeed(page: Page) {
 
     await route.fulfill({
       status: 200,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify(payload),
     });
   });
@@ -427,34 +375,7 @@ async function registerMockBookingsFeed(page: Page) {
     }
   }
 
-  // Intercept Object.defineProperty BEFORE webpack loads so that
-  // target exports are created as configurable (allowing later override)
-  await page.addInitScript(() => {
-    const targetExports = new Set([
-      "clientGetDataByCalendarEventId",
-      "clientFetchAllDataFromCollection",
-      "getPaginatedData",
-    ]);
-    const origDefineProperty = Object.defineProperty;
-    Object.defineProperty = function (
-      obj: any,
-      prop: PropertyKey,
-      descriptor: PropertyDescriptor
-    ) {
-      if (
-        descriptor &&
-        descriptor.get &&
-        !descriptor.configurable &&
-        typeof prop === "string" &&
-        targetExports.has(prop)
-      ) {
-        descriptor = { ...descriptor, configurable: true };
-      }
-      return origDefineProperty.call(this, obj, prop, descriptor);
-    } as typeof Object.defineProperty;
-    // Preserve other static methods
-    Object.defineProperty.toString = () => "function defineProperty() { [native code] }";
-  });
+  await registerDefinePropertyInterceptor(page);
 
   await page.addInitScript(
     ({ timestampFields, bookings, serviceEmail }) => {
@@ -542,7 +463,8 @@ async function registerMockBookingsFeed(page: Page) {
 
         if (
           normalizedTableName.includes("booking") &&
-          !normalizedTableName.includes("type")
+          !normalizedTableName.includes("type") &&
+          !normalizedTableName.includes("log")
         ) {
           return await ensureBookings();
         }
@@ -568,7 +490,8 @@ async function registerMockBookingsFeed(page: Page) {
 
         if (
           normalizedTableName.includes("booking") &&
-          !normalizedTableName.includes("type")
+          !normalizedTableName.includes("type") &&
+          !normalizedTableName.includes("log")
         ) {
           return await ensureBookings();
         }
@@ -712,11 +635,11 @@ async function registerMockBookingsFeed(page: Page) {
           safetyTrainedUsers: [],
         };
       } else {
-        loadBookings();
+        void loadBookings().catch(() => {});
       }
     },
     {
-      timestampFields: TIMESTAMP_FIELDS,
+      timestampFields: TIMESTAMP_FIELDS_WITH_BY,
       bookings: initialPayload,
       serviceEmail: SERVICE_USER_EMAIL,
     }
@@ -790,7 +713,7 @@ async function mockServicesEndpoint(page: Page, captured: any[]) {
     if (route.request().method() !== "POST") {
       await route.fulfill({
         status: 405,
-        headers: jsonHeaders,
+        headers: JSON_HEADERS,
         body: JSON.stringify({ error: "Method Not Allowed" }),
       });
       return;
@@ -818,7 +741,7 @@ async function mockServicesEndpoint(page: Page, captured: any[]) {
 
     await route.fulfill({
       status: 200,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify({ success: true }),
     });
   });
@@ -829,14 +752,14 @@ async function mockTransitionEndpoints(page: Page) {
     if (route.request().method() === "POST") {
       await route.fulfill({
         status: 200,
-        headers: jsonHeaders,
+        headers: JSON_HEADERS,
         body: JSON.stringify({ success: true }),
       });
       return;
     }
     await route.fulfill({
       status: 405,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify({ error: "Method Not Allowed" }),
     });
   });
@@ -845,14 +768,14 @@ async function mockTransitionEndpoints(page: Page) {
     if (route.request().method() === "POST") {
       await route.fulfill({
         status: 200,
-        headers: jsonHeaders,
+        headers: JSON_HEADERS,
         body: JSON.stringify({ success: true }),
       });
       return;
     }
     await route.fulfill({
       status: 200,
-      headers: jsonHeaders,
+      headers: JSON_HEADERS,
       body: JSON.stringify([]),
     });
   });
